+++ src/measures/NetGO.cpp
@@ -78,8 +78,8 @@ double NetGO::eval(const Alignment& A) {
     unordered_map<uint,uint> goCountG1 = GoSimilarity::getGoCounts(*G1);
     unordered_map<uint,uint> goCountG2 = GoSimilarity::getGoCounts(*G2);
 #if 0
-    map<ushort,string> namesG1 = G1->getIndexToNodeNameMap();
-    map<ushort,string> namesG2 = G2->getIndexToNodeNameMap();
+    map<uint,string> namesG1 = G1->getIndexToNodeNameMap();
+    map<uint,string> namesG2 = G2->getIndexToNodeNameMap();
 #endif
     double total = 0;
     for (uint i = 0; i < n1; i++) {
