+++ src/measures/TriangleCorrectness.cpp
@@ -21,7 +21,7 @@ double TriangleCorrectness::eval(const Alignment& A){
 
 int TriangleCorrectness::calculateTriangles(Graph* G){
     int numTriangles = 0;
-    vector<vector<ushort> > GAdjLists;
+    vector<vector<uint> > GAdjLists;
     Matrix GMatrix;
     G->getAdjLists(GAdjLists);
     G->getMatrix(GMatrix);
@@ -29,8 +29,8 @@ int TriangleCorrectness::calculateTriangles(Graph* G){
         for(uint j = 0; j < GAdjLists[i].size(); j++){
             for(uint k = 0; k < GAdjLists[i].size(); k++){
                 if(k != j){
-                    ushort neighbor1 = GAdjLists[i][j];
-                    ushort neighbor2 = GAdjLists[i][k];
+                    uint neighbor1 = GAdjLists[i][j];
+                    uint neighbor2 = GAdjLists[i][k];
                     if(GMatrix.get(neighbor1, neighbor2)){
                         numTriangles++;
                     }
